change getArmyInfo request over to store fn fetchArmyDetails
