+++ src/pages/ArmyInfo/ArmyInfo.tsx
@@ -1,28 +1,15 @@
 import { useLocation } from "react-router-dom";
 import "./ArmyInfo.scss";
 import { useEffect, useState } from "react";
-import { getArmyInfo } from "../../utils/ArmyRequests";
 import ArmyDash from "../../components/ArmyDash/ArmyDash";
 import ArmyNemesis from "../../components/ArmyNemesis/ArmyNemesis";
 import ArmyAlly from "../../components/ArmyAlly/ArmyAlly";
 import { Player } from "../../utils/Interfaces";
 import { CircularProgress } from "@mui/material";
 import RankGraph from "../../components/RankGraph/RankGraph";
-
-export interface ArmyInformation {
-  id: string;
-  ranking: string;
-  user_id: string;
-  emblem_id: string;
-  rn: number;
-  type: "fantasy" | "40k";
-  name: string;
-  emblem: string;
-  known_as: string;
-}
+import { useArmiesStore } from "../../store/armies";
 
 export default function ArmyInfo() {
-  const [armyObj, setArmyObj] = useState<ArmyInformation>();
   const [battleCount, setBattleCount] = useState(0);
   const [winPercent, setWinPercent] = useState("");
   const [armyRank, setArmyRank] = useState(0);
@@ -30,28 +17,27 @@ export default function ArmyInfo() {
   const [allyObj, setAllyObj] = useState<Player | null>(null);
 
   const location = useLocation();
+  const { fetchArmyDetails, selectedArmy } = useArmiesStore();
 
   const armyID = location.state.id;
 
   useEffect(() => {
     const fetchData = async () => {
-      if (!armyObj) {
-        const response = await getArmyInfo(armyID, 3);
-        if (response) {
-          setArmyObj(response.user);
-          setBattleCount(response.battleCount);
-          setWinPercent(response.winPercent);
-          setArmyRank(Number(response.user.ranking));
-          setNemesisObj(response.nemesis);
-          setAllyObj(response.ally);
-        }
+      if (!selectedArmy) {
+        await fetchArmyDetails(armyID);
+      } else {
+        setBattleCount(selectedArmy.battleCount);
+        setWinPercent(selectedArmy.winPercent);
+        setArmyRank(Number(selectedArmy.user.ranking));
+        setNemesisObj(selectedArmy.nemesis);
+        setAllyObj(selectedArmy.ally);
       }
     };
 
     fetchData();
-  }, [armyID]);
+  }, [armyID, selectedArmy, fetchArmyDetails]);
 
-  if (!armyObj) {
+  if (!selectedArmy) {
     return (
       <div className="loading-message">
         <CircularProgress style={{ color: "white" }} />
@@ -64,10 +50,10 @@ export default function ArmyInfo() {
       <ArmyDash
         winPercent={winPercent}
         battleCount={battleCount}
-        armyObj={armyObj}
+        armyObj={selectedArmy.user}
         armyRank={armyRank}
       />
-      <RankGraph army_id={armyObj.id} name={armyObj.name} />
+      <RankGraph army_id={selectedArmy.user.id} name={selectedArmy.user.name} />
       <section className="army-info__nemesis-ally">
         <div className="army-info__container">
           <ArmyNemesis nemesis={nemesisObj} />
